requirements.txt
Django==5.0
djangorestframework==3.14.0
Pillow==10.1.0
qrcode==7.4.2
django-cors-headers==4.3.1
python-decouple==3.8